move log-method to mylanguages to use the initialized names, etc.
